requirements.txt
dash
dash_bootstrap_components
dash_bootstrap_templates
Flask
genapy
python-nmap
numpy
pandas
plotly
pyserial
PyYAML
requests
scipy
tailer
python-dotenv
tensorboardX==2.6.2.2
tabulate